+++ packages/docs/angular-docs/src/components/layout/app-docs-layout.tsx
@@ -93,7 +93,7 @@ export function AppDocsLayout({children, ...props}: Props): ReactNode {
               <ResetDemoDimensionsButton />
               <HeaderRepositoryLink />
               <GlobalConfig />
-              <HeaderAiLink modelId="cs45-qui-angular-nextgen" />
+              <HeaderAiLink modelId="qui-angular-nextgen-sonnet-4-5" />
               <ThemeToggle />
             </HeaderBar.ActionBar>
           </HeaderBar.Root>