+++ packages/frameworks/angular-core/src/side-nav/index.ts
@@ -1,1 +1,7 @@
-export {}
+export * from "./core-side-nav-header-action.directive"
+export * from "./core-side-nav-header-logo.directive"
+export * from "./core-side-nav-header-title.directive"
+export * from "./core-side-nav-header.directive"
+export * from "./core-side-nav-root.directive"
+export * from "./core-side-nav-trigger.directive"
+export * from "./side-nav-context.service"