+++ packages/frameworks/angular/src/side-nav/side-nav-branch-indent-guide.directive.ts
@@ -1,0 +1,23 @@
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import {computed, Directive} from "@angular/core"
+
+import {CoreTreeBranchIndentGuideDirective} from "@qualcomm-ui/angular-core/tree"
+
+import {useQdsSideNavContext} from "./qds-side-nav-context.service"
+
+@Directive({
+  selector: "[q-side-nav-branch-indent-guide]",
+  standalone: false,
+})
+export class SideNavBranchIndentGuideDirective extends CoreTreeBranchIndentGuideDirective {
+  protected qdsContext = useQdsSideNavContext()
+
+  constructor() {
+    super()
+    this.trackBindings.extendWith(
+      computed(() => this.qdsContext().getBranchIndentGuideBindings()),
+    )
+  }
+}