+++ packages/frameworks/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qualcomm-ui/react
 
+## 1.4.2 (2025/12/03)
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/qds-core]
+
 ## 1.4.1 (2025/12/03)
 
 ### Miscellaneous Chores